refactoring SortUtil with method template pattern
