+++ src/main/java/com/idforideas/pizzeria/appuser/AppUserResource.java
@@ -25,7 +25,7 @@ import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.idforideas.pizzeria.utils.Response;
+import com.idforideas.pizzeria.util.Response;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
