+++ src/main/java/com/idforideas/pizzeria/util/SortArrays.java
@@ -1,22 +1,19 @@
-package com.idforideas.pizzeria.utils;
+package com.idforideas.pizzeria.util;
 
 import static java.util.Arrays.stream;
-import static org.springframework.data.domain.Sort.Direction.ASC;
-import static org.springframework.data.domain.Sort.Direction.DESC;
-import static org.springframework.data.domain.Sort.Order.desc;
 import static org.springframework.data.domain.Sort.Order.by;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.ArrayList;
 import java.util.stream.IntStream;
 
 import com.idforideas.pizzeria.exceptions.BadRequestException;
 
 import org.springframework.data.domain.Sort.Order;
 
-public abstract class SortUtil {
+public class SortArrays extends SortUtil {
 
-   public static List<Order> getOrders(String[] sort) {
+    public List<Order> getOrders(String[] sort) {
         final List<Order> orders = new ArrayList<>();
         final boolean withDirection = stream(sort).anyMatch(SortUtil::isDirection);
 
@@ -36,9 +33,6 @@ public abstract class SortUtil {
         return orders;
     }
 
-    private static Order getOrder(String property, String direction) {
-        return direction.equalsIgnoreCase(ASC.name()) ? by(property) : desc(property);
-    }
 
     private static boolean isValid(String[] sort) {
         if(hasDirectionAtZeroIndex(sort[0])) return false;
@@ -68,9 +62,5 @@ public abstract class SortUtil {
         }
 
         return isTogether;
-    }
-
-    private static boolean isDirection(String arg) {
-        return arg.equalsIgnoreCase(ASC.name()) || arg.equalsIgnoreCase(DESC.name());
-    }
+    } 
 }
